Lock wingEdge in triangleNeighbors
It got stuck with one thread running in an infinite loop and the other
seven paused. It was stuck in triangleNeighbors. The reason appeared
to be that another thread had flipped an edge while this one was
following edge pointers around a point, and it never came back
to the starting edge pointer.