Attempt to prevent a crash
The crash occurred when the thread's main function stepped the edge index
to the last edge, which had just been added. It tried to check whether
the edge should be bent when neither of the edge's triangle pointers
had been set, resulting in a call to shouldSplit(0x0,0.2). While
shouldSplit was running, the thread that added the edge set both triangle
pointers, making it an interior edge, about which shouldBend should
not have been asked.